tknewies

+++ ft_memcpy.c
@@ -6,22 +6,22 @@
 /*   By: svoort <svoort@student.codam.nl>             +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2019/01/09 11:23:58 by svoort         #+#    #+#                */
-/*   Updated: 2019/01/14 14:27:59 by svoort        ########   odam.nl         */
+/*   Updated: 2019/01/18 10:28:06 by svoort        ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "./includes/libft.h"
+#include "libft.h"
 
 void	*ft_memcpy(void *dst, const void *src, size_t n)
 {
-	int		i;
+	size_t	i;
 	char	*cdst;
 	char	*csrc;
 
 	i = 0;
 	cdst = (char*)dst;
 	csrc = (char*)src;
-	while (i < (int)n)
+	while (i < n)
 	{
 		cdst[i] = csrc[i];
 		i++;
